refactor: :recycle: Refactored solution
- Implemented BusinessDayTool for business day calculations.
- Created CalculateDifferenceTool for calculating time differences between two dates.
- Developed ConvertTimezoneTool for converting dates between timezones.
- Added DateQueryTool for advanced date queries and navigation.
- Introduced FormatDurationTool for formatting duration values.
- Implemented GetCurrentDateTool for retrieving current date and time with timezone support.
- Created SubtractTimeTool for subtracting time from dates.
- Added utility functions for business day operations and date queries.
- Defined type interfaces for command options and parameters.
- Implemented utility functions for date formatting, parsing, and duration formatting.
