+++ src/registration/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Registration modules for AI Watch extension.
+ */
+
+export * from './commands';
+export * from './tools';
