+++ src/registration/tools.ts
@@ -1,0 +1,42 @@
+/**
+ * Language Model Tools registration for AI Watch extension.
+ */
+
+import * as vscode from 'vscode';
+import {
+  GetCurrentDateTool,
+  CalculateDifferenceTool,
+  ConvertTimezoneTool,
+  AddTimeTool,
+  SubtractTimeTool,
+  FormatDurationTool,
+  BusinessDayTool,
+  DateQueryTool,
+} from '../tools';
+
+/**
+ * Registers all Language Model Tools with VS Code for AI assistant integration.
+ * These tools enable AI assistants to perform sophisticated date/time operations.
+ *
+ * @param context - VS Code extension context for managing subscriptions
+ */
+export function registerChatTools(context: vscode.ExtensionContext) {
+  context.subscriptions.push(
+    vscode.lm.registerTool('ai-watch_getCurrentDate', new GetCurrentDateTool()),
+  );
+  context.subscriptions.push(
+    vscode.lm.registerTool('ai-watch_calculateDifference', new CalculateDifferenceTool()),
+  );
+  context.subscriptions.push(
+    vscode.lm.registerTool('ai-watch_convertTimezone', new ConvertTimezoneTool()),
+  );
+  context.subscriptions.push(vscode.lm.registerTool('ai-watch_addTime', new AddTimeTool()));
+  context.subscriptions.push(
+    vscode.lm.registerTool('ai-watch_subtractTime', new SubtractTimeTool()),
+  );
+  context.subscriptions.push(
+    vscode.lm.registerTool('ai-watch_formatDuration', new FormatDurationTool()),
+  );
+  context.subscriptions.push(vscode.lm.registerTool('ai-watch_businessDay', new BusinessDayTool()));
+  context.subscriptions.push(vscode.lm.registerTool('ai-watch_dateQuery', new DateQueryTool()));
+}
