+++ src/types/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Type definitions for AI Watch extension.
+ */
+
+export * from './options';
+export * from './parameters';
